Extract RacingAxisTop
